Passive Separation (#829)
## Misc. Changes
- Moves methods to their respective passive classes.
- Renames Hydrosink,java to HydroSink.java
## Removals
- Removes WaterPassive.java
- Removes AirPassive.java
