+++ src/com/projectkorra/projectkorra/PKListener.java
@@ -176,12 +176,10 @@ import com.projectkorra.projectkorra.waterbending.ice.IceSpikeBlast;
 import com.projectkorra.projectkorra.waterbending.ice.PhaseChange;
 import com.projectkorra.projectkorra.waterbending.ice.PhaseChange.PhaseChangeType;
 import com.projectkorra.projectkorra.waterbending.multiabilities.WaterArms;
-import com.projectkorra.projectkorra.waterbending.passive.Hydrosink;
-import com.projectkorra.projectkorra.waterbending.passive.WaterPassive;
+import com.projectkorra.projectkorra.waterbending.passive.HydroSink;
 import com.projectkorra.rpg.RPGMethods;
 
 public class PKListener implements Listener {
-
 	ProjectKorra plugin;
 
 	private static final HashMap<Entity, Ability> BENDING_ENTITY_DEATH = new HashMap<>(); // Entities killed by Bending.
@@ -196,8 +194,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onBlockBreak(BlockBreakEvent event) {
-
-
 		Block block = event.getBlock();
 		Player player = event.getPlayer();
 		BendingPlayer bPlayer = BendingPlayer.getBendingPlayer(player);
@@ -240,24 +236,25 @@ public class PKListener implements Listener {
 			EarthAbility.removeRevertIndex(block);
 		} else if (TempBlock.isTempBlock(block)) {
 			TempBlock.revertBlock(block, Material.AIR);
-		} else if (EarthPassive.isPassiveSand(block)) {
-			EarthPassive.revertSand(block);
+		} else if (DensityShift.isPassiveSand(block)) {
+			DensityShift.revertSand(block);
 		}
 	}
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onBlockFlowTo(BlockFromToEvent event) {
-		
 		Block toblock = event.getToBlock();
 		Block fromblock = event.getBlock();
 		if (ElementalAbility.isLava(fromblock)) {
 			event.setCancelled(!EarthPassive.canFlowFromTo(fromblock, toblock));
 		}
+		
 		if (ElementalAbility.isWater(fromblock)) {
 			event.setCancelled(!WaterBubble.isAir(toblock));
 			if (!event.isCancelled()) {
 				event.setCancelled(!WaterManipulation.canFlowFromTo(fromblock, toblock));
 			}
+			
 			if (!event.isCancelled()) {
 				if (TempBlock.isTempBlock(toblock) && Illumination.getBlocks().containsKey(TempBlock.get(toblock))) {
 					toblock.setType(Material.AIR);
@@ -268,16 +265,18 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onBlockForm(BlockFormEvent event) {
-		
 		if (TempBlock.isTempBlock(event.getBlock())) {
 			event.setCancelled(true);
 		}
+		
 		if (!WaterManipulation.canPhysicsChange(event.getBlock())) {
 			event.setCancelled(true);
 		}
+		
 		if (!EarthPassive.canPhysicsChange(event.getBlock())) {
 			event.setCancelled(true);
 		}
+		
 		if (event.getBlock().getType().toString().equals("CONCRETE_POWDER")) {
 			BlockFace[] faces = new BlockFace[] {BlockFace.UP, BlockFace.DOWN, BlockFace.EAST, BlockFace.WEST, BlockFace.NORTH, BlockFace.SOUTH};
 			
@@ -299,33 +298,36 @@ public class PKListener implements Listener {
 	}
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
-	public void onBlockIgnite(BlockIgniteEvent event) {
-		
-	}
+	public void onBlockIgnite(BlockIgniteEvent event) {}
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onBlockMeltEvent(BlockFadeEvent event) {
-		
 		Block block = event.getBlock();
 		if (block.getType() == Material.FIRE) {
 			return;
 		}
+		
 		event.setCancelled(TempBlock.isTempBlock(block) && Illumination.getBlocks().containsKey(TempBlock.get(block)));
 		if (!event.isCancelled()) {
 			event.setCancelled(!WaterManipulation.canPhysicsChange(block));
 		}
+		
 		if (!event.isCancelled()) {
 			event.setCancelled(!EarthPassive.canPhysicsChange(block));
 		}
+		
 		if (!event.isCancelled()) {
 			event.setCancelled(PhaseChange.getFrozenBlocksAsBlock().contains(block));
 		}
+		
 		if (!event.isCancelled()) {
 			event.setCancelled(!SurgeWave.canThaw(block));
 		}
+		
 		if (!event.isCancelled()) {
 			event.setCancelled(!Torrent.canThaw(block));
 		}
+		
 		if (BlazeArc.getIgnitedBlocks().containsKey(block)) {
 			BlazeArc.removeBlock(block);
 		}
@@ -333,7 +335,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onBlockPhysics(BlockPhysicsEvent event) {
-		
 		Block block = event.getBlock();
 
 		if (!WaterManipulation.canPhysicsChange(block) || !EarthPassive.canPhysicsChange(block) || (TempBlock.isTempBlock(block) && Illumination.getBlocks().containsKey(TempBlock.get(block))) || EarthAbility.getPreventPhysicsBlocks().contains(block)) {
@@ -348,7 +349,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onBlockPlace(BlockPlaceEvent event) {
-		
 		Player player = event.getPlayer();
 		if (MovementHandler.isStopped(player) || Bloodbending.isBloodbent(player) || Suffocate.isBreathbent(player)) {
 			event.setCancelled(true);
@@ -376,6 +376,7 @@ public class PKListener implements Listener {
 			} else {
 				prefix = ChatColor.WHITE + ChatColor.translateAlternateColorCodes('&', ConfigManager.languageConfig.get().getString("Chat.Prefixes.Nonbender")) + " ";
 			}
+			
 			player.setDisplayName(player.getName());
 			player.setDisplayName(prefix + ChatColor.RESET + player.getDisplayName());
 		}
@@ -389,8 +390,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onEntityChangeBlockEvent(EntityChangeBlockEvent event) {
-		
-
 		Entity entity = event.getEntity();
 		if (MovementHandler.isStopped(entity) || Bloodbending.isBloodbent(entity) || Suffocate.isBreathbent(entity)) {
 			event.setCancelled(true);
@@ -406,8 +405,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onEntityCombust(EntityCombustEvent event) {
-		
-
 		Entity entity = event.getEntity();
 		Block block = entity.getLocation().getBlock();
 		if (BlazeArc.getIgnitedBlocks().containsKey(block) && entity instanceof LivingEntity) {
@@ -416,9 +413,7 @@ public class PKListener implements Listener {
 	}
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
-	public void onEntityDamageBlock(EntityDamageByBlockEvent event) {
-
-	}
+	public void onEntityDamageBlock(EntityDamageByBlockEvent event) {}
 
 	@EventHandler(priority = EventPriority.LOWEST, ignoreCancelled = true)
 	public void onEntityDamageByBlock(EntityDamageByBlockEvent event) {
@@ -436,8 +431,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onEntityDamageEvent(EntityDamageEvent event) {
-		
-
 		Entity entity = event.getEntity();
 
 		if (event.getCause() == DamageCause.FIRE && BlazeArc.getIgnitedBlocks().containsKey(entity.getLocation().getBlock())) {
@@ -465,6 +458,7 @@ public class PKListener implements Listener {
 			if (bPlayer.isElementToggled(Element.FIRE)) {
 				return;
 			}
+			
 			if (bPlayer.getBoundAbilityName().equalsIgnoreCase("HeatControl")) {
 				if (event.getCause() == DamageCause.FIRE || event.getCause() == DamageCause.FIRE_TICK) {
 					player.setFireTicks(0);
@@ -542,23 +536,25 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.HIGH, ignoreCancelled = true)
 	public void onEntityExplode(EntityExplodeEvent event) {
-		
-
 		for (Block block : event.blockList()) {
 			EarthBlast blast = EarthBlast.getBlastFromSource(block);
 
 			if (blast != null) {
 				blast.remove();
 			}
+			
 			if (PhaseChange.getFrozenBlocksAsBlock().contains(block)) {
 				PhaseChange.thaw(block);
 			}
+			
 			if (SurgeWall.getWallBlocks().containsKey(block)) {
 				block.setType(Material.AIR);
 			}
+			
 			if (!SurgeWave.canThaw(block)) {
 				SurgeWave.thaw(block);
 			}
+			
 			if (EarthAbility.getMovedEarth().containsKey(block)) {
 				EarthAbility.removeRevertIndex(block);
 			}
@@ -567,8 +563,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onEntityExplodeEvent(EntityExplodeEvent event) {
-		
-
 		Entity entity = event.getEntity();
 		if (entity != null) {
 			if (MovementHandler.isStopped(entity) || Bloodbending.isBloodbent(entity) || Suffocate.isBreathbent(entity)) {
@@ -579,8 +573,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onEntityInteractEvent(EntityInteractEvent event) {
-		
-
 		Entity entity = event.getEntity();
 		if (MovementHandler.isStopped(entity) || Bloodbending.isBloodbent(entity) || Suffocate.isBreathbent(entity)) {
 			event.setCancelled(true);
@@ -589,8 +581,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onEntityProjectileLaunchEvent(ProjectileLaunchEvent event) {
-		
-
 		Entity entity = event.getEntity();
 		if (MovementHandler.isStopped(entity) || Bloodbending.isBloodbent(entity) || Suffocate.isBreathbent(entity)) {
 			event.setCancelled(true);
@@ -599,8 +589,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onEntityShootBowEvent(EntityShootBowEvent event) {
-		
-
 		Entity entity = event.getEntity();
 		if (MovementHandler.isStopped(entity) || Bloodbending.isBloodbent(entity) || Suffocate.isBreathbent(entity)) {
 			event.setCancelled(true);
@@ -609,8 +597,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onEntitySlimeSplitEvent(SlimeSplitEvent event) {
-		
-
 		Entity entity = event.getEntity();
 		if (MovementHandler.isStopped(entity) || Bloodbending.isBloodbent(entity) || Suffocate.isBreathbent(entity)) {
 			event.setCancelled(true);
@@ -619,8 +605,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onEntitySuffocatedByTempBlocks(EntityDamageEvent event) {
-		
-
 		if (event.getCause() == DamageCause.SUFFOCATION) {
 			if (TempBlock.isTempBlock(event.getEntity().getLocation().add(0, 1, 0).getBlock())) {
 				event.setCancelled(true);
@@ -630,8 +614,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onEntityTarget(EntityTargetEvent event) {
-		
-
 		Entity entity = event.getEntity();
 		if (MovementHandler.isStopped(entity) || Bloodbending.isBloodbent(entity)) {
 			event.setCancelled(true);
@@ -640,7 +622,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onEntityTargetLiving(EntityTargetLivingEntityEvent event) {
-		
 		Entity entity = event.getEntity();
 		if (MovementHandler.isStopped(entity) || Bloodbending.isBloodbent(entity)) {
 			event.setCancelled(true);
@@ -649,8 +630,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onEntityTeleportEvent(EntityTeleportEvent event) {
-		
-
 		Entity entity = event.getEntity();
 		if (MovementHandler.isStopped(entity) || Bloodbending.isBloodbent(entity) || Suffocate.isBreathbent(entity) || MetalClips.isControlled((LivingEntity)entity)) {
 			event.setCancelled(true);
@@ -689,8 +668,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onInventoryClick(InventoryClickEvent event) {
-		
-
 		for (MetalClips clips : CoreAbility.getAbilities(MetalClips.class)) {
 			if (clips.getTargetEntity() != null && clips.getTargetEntity().getEntityId() == event.getWhoClicked().getEntityId()) {
 				event.setCancelled(true);
@@ -710,7 +687,6 @@ public class PKListener implements Listener {
 	@EventHandler(priority = EventPriority.NORMAL)
 	public void onEntityBendingDeath(EntityBendingDeathEvent event) {
 		BENDING_ENTITY_DEATH.put(event.getEntity(), event.getAbility());
-		
 		if (ConfigManager.languageConfig.get().getBoolean("DeathMessages.Enabled") && event.getEntity() instanceof Player) {
 			Ability ability = event.getAbility();
 			if (ability == null) {
@@ -731,8 +707,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onPlayerChat(AsyncPlayerChatEvent event) {
-		
-
 		Player player = event.getPlayer();
 		BendingPlayer bPlayer = BendingPlayer.getBendingPlayer(player);
 
@@ -770,7 +744,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onPlayerDamage(EntityDamageEvent event) {
-
 		if (event.getEntity() instanceof Player) {
 			Player player = (Player) event.getEntity();
 			BendingPlayer bPlayer = BendingPlayer.getBendingPlayer(player);
@@ -796,15 +769,15 @@ public class PKListener implements Listener {
 				}
 			}
 
-			if (bPlayer.hasElement(Element.WATER) && event.getCause() == DamageCause.FALL && bPlayer.canBendPassive(Element.WATER) && bPlayer.canUsePassive(Element.WATER) && CoreAbility.getAbility(Hydrosink.class).isEnabled() && PassiveManager.hasPassive(player, CoreAbility.getAbility(Hydrosink.class))) {
-				if (WaterPassive.applyNoFall(player)) {
+			if (bPlayer.hasElement(Element.WATER) && event.getCause() == DamageCause.FALL && bPlayer.canBendPassive(Element.WATER) && bPlayer.canUsePassive(Element.WATER) && CoreAbility.getAbility(HydroSink.class).isEnabled() && PassiveManager.hasPassive(player, CoreAbility.getAbility(HydroSink.class))) {
+				if (HydroSink.applyNoFall(player)) {
 					event.setDamage(0D);
 					event.setCancelled(true);
 				}
 			}
 
 			if (bPlayer.hasElement(Element.EARTH) && event.getCause() == DamageCause.FALL && bPlayer.canBendPassive(Element.EARTH) && bPlayer.canUsePassive(Element.EARTH) && CoreAbility.getAbility(DensityShift.class).isEnabled() && PassiveManager.hasPassive(player, CoreAbility.getAbility(DensityShift.class))) {
-				if (EarthPassive.softenLanding(player)) {
+				if (DensityShift.softenLanding(player)) {
 					event.setDamage(0D);
 					event.setCancelled(true);
 				}
@@ -812,7 +785,7 @@ public class PKListener implements Listener {
 
 			if (bPlayer.hasElement(Element.CHI) && event.getCause() == DamageCause.FALL && bPlayer.canBendPassive(Element.CHI) && bPlayer.canUsePassive(Element.CHI) && CoreAbility.getAbility(Acrobatics.class).isEnabled() && PassiveManager.hasPassive(player, CoreAbility.getAbility(Acrobatics.class))) {
 				double initdamage = event.getDamage();
-				double newdamage = event.getDamage() * ChiPassive.getFallReductionFactor();
+				double newdamage = event.getDamage() * Acrobatics.getFallReductionFactor();
 				double finaldamage = initdamage - newdamage;
 				event.setDamage(finaldamage);
 				if (finaldamage <= 0.4)
@@ -844,8 +817,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onPlayerDamageByPlayer(EntityDamageByEntityEvent e) {
-		
-
 		Entity source = e.getDamager();
 		Entity entity = e.getEntity();
 		FireBlastCharged fireball = FireBlastCharged.getFireball(source);
@@ -942,6 +913,7 @@ public class PKListener implements Listener {
 				if (coreAbil != null) {
 					element = coreAbil.getElement();
 				}
+				
 				if (HorizontalVelocityTracker.hasBeenDamagedByHorizontalVelocity(event.getEntity()) && Arrays.asList(HorizontalVelocityTracker.abils).contains(tempAbility)) {
 					if (ConfigManager.languageConfig.get().contains("Abilities." + element.getName() + "." + tempAbility + ".HorizontalVelocityDeath")) {
 						message = ConfigManager.languageConfig.get().getString("Abilities." + element.getName() + "." + tempAbility + ".HorizontalVelocityDeath");
@@ -964,6 +936,7 @@ public class PKListener implements Listener {
 						message = ConfigManager.languageConfig.get().getString("Abilities.Avatar.Combo." + tempAbility + ".DeathMessage");
 					}
 				}
+				
 				message = message.replace("{victim}", event.getEntity().getName()).replace("{attacker}", event.getEntity().getKiller().getName()).replace("{ability}", ability);
 				event.setDeathMessage(message);
 				BENDING_PLAYER_DEATH.remove(event.getEntity());
@@ -973,7 +946,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onPlayerInteraction(PlayerInteractEvent event) {
-		
 		Player player = event.getPlayer();
 		BendingPlayer bPlayer = BendingPlayer.getBendingPlayer(player);
 
@@ -1004,6 +976,7 @@ public class PKListener implements Listener {
 				new EarthSmash(player, ClickType.RIGHT_CLICK);
 			}
 		}
+		
 		if (MovementHandler.isStopped(player) || Bloodbending.isBloodbent(player) || Suffocate.isBreathbent(player)) {
 			event.setCancelled(true);
 		}
@@ -1011,8 +984,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onPlayerInteractEntity(PlayerInteractEntityEvent event) {
-		
-
 		Player player = event.getPlayer();
 		BendingPlayer bPlayer = BendingPlayer.getBendingPlayer(player);
 
@@ -1035,8 +1006,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onPlayerItemDamage(PlayerItemDamageEvent event) {
-		
-
 		if (TempArmor.hasTempArmor(event.getPlayer())) {
 			TempArmor armor = TempArmor.getTempArmor(event.getPlayer());
 			for (ItemStack i : armor.getNewArmor()) {
@@ -1089,16 +1058,12 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onPlayerKick(PlayerKickEvent event) {
-		
-
 		AirFlight.remove(event.getPlayer());
 		JUMPS.remove(event.getPlayer());
 	}
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onPlayerMove(PlayerMoveEvent event) {
-		
-
 		Player player = event.getPlayer();
 		BendingPlayer bPlayer = BendingPlayer.getBendingPlayer(player);
 		if (MovementHandler.isStopped(player)) {
@@ -1106,9 +1071,7 @@ public class PKListener implements Listener {
 				event.setCancelled(true);
 				return;
 			}
-		}
-
-		else if (CoreAbility.hasAbility(player, WaterSpout.class) || CoreAbility.hasAbility(player, AirSpout.class) ) {
+		} else if (CoreAbility.hasAbility(player, WaterSpout.class) || CoreAbility.hasAbility(player, AirSpout.class) ) {
 			Vector vel = new Vector();
 			vel.setX(event.getTo().getX() - event.getFrom().getX());
 			vel.setZ(event.getTo().getZ() - event.getFrom().getZ());
@@ -1121,14 +1084,13 @@ public class PKListener implements Listener {
 				// apply the new velocity
 				event.getPlayer().setVelocity(vel);
 			}
-		}
-
-		else if (Bloodbending.isBloodbent(player)) {
+		} else if (Bloodbending.isBloodbent(player)) {
 			BendingPlayer bender = Bloodbending.getBloodbender(player);
 			if (bender.isAvatarState()) {
 				event.setCancelled(true);
 				return;
 			}
+			
 			double distance1 = 0;
 			double distance2 = 0;
 			Location loc = Bloodbending.getBloodbendingLocation(player);
@@ -1145,9 +1107,7 @@ public class PKListener implements Listener {
 					return;
 				}
 			}
-		}
-
-		else if (AirFlight.isFlying(player)) {
+		} else if (AirFlight.isFlying(player)) {
 			if (AirFlight.isHovering(player)) {
 				Location loc = event.getFrom();
 				Location toLoc = event.getTo();
@@ -1157,9 +1117,7 @@ public class PKListener implements Listener {
 					return;
 				}
 			}
-		}
-
-		else {
+		} else {
 			if (bPlayer != null) {
 				if (bPlayer.hasElement(Element.AIR) || bPlayer.hasElement(Element.CHI)) {
 					if (event.getTo().getX() != event.getFrom().getX() || event.getTo().getY() != event.getFrom().getY() || event.getTo().getZ() != event.getFrom().getZ()) {
@@ -1209,6 +1167,7 @@ public class PKListener implements Listener {
 			if (TOGGLED_OUT.contains(player.getUniqueId()) && bPlayer.isToggled()) {
 				TOGGLED_OUT.remove(player.getUniqueId());
 			}
+			
 			if (!bPlayer.isToggled()) {
 				TOGGLED_OUT.add(player.getUniqueId());
 			}
@@ -1217,6 +1176,7 @@ public class PKListener implements Listener {
 		if (Commands.invincible.contains(player.getName())) {
 			Commands.invincible.remove(player.getName());
 		}
+		
 		Preset.unloadPreset(player);
 
 		if (TempArmor.hasTempArmor(player)) {
@@ -1278,6 +1238,7 @@ public class PKListener implements Listener {
 						if (block.getData() >= 8) {
 							block = block.getRelative(BlockFace.DOWN);
 						}
+						
 						if (block.getData() < 4) {
 							block.setData((byte) (block.getData() + 4));
 							block.getWorld().playSound(block.getLocation(), Sound.BLOCK_IRON_DOOR_CLOSE, 10, 1);
@@ -1285,6 +1246,7 @@ public class PKListener implements Listener {
 							block.setData((byte) (block.getData() - 4));
 							block.getWorld().playSound(block.getLocation(), Sound.BLOCK_IRON_DOOR_OPEN, 10, 1);
 						}
+						
 						bPlayer.addCooldown("FerroControl", 200);
 					}
 				}
@@ -1309,6 +1271,7 @@ public class PKListener implements Listener {
 			if (coreAbil instanceof AddonAbility) {
 				return;
 			}
+			
 			if (coreAbil instanceof AirAbility && bPlayer.isElementToggled(Element.AIR) == true) {
 				if (bPlayer.canCurrentlyBendWithWeapons()) {
 					if (abil.equalsIgnoreCase("Tornado")) {
@@ -1329,6 +1292,7 @@ public class PKListener implements Listener {
 						if (player.isSneaking() || !bPlayer.canUseFlight()) {
 							return;
 						}
+						
 						new AirFlight(player);
 					}
 				}
@@ -1362,6 +1326,7 @@ public class PKListener implements Listener {
 					} else if (abil.equalsIgnoreCase("WaterArms")) {
 						new WaterArms(player);
 					}
+					
 					if (abil.equalsIgnoreCase("HealingWaters")) {
 						new HealingWaters(player);
 					}
@@ -1453,8 +1418,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onPlayerSwing(PlayerAnimationEvent event) {
-		
-
 		Player player = event.getPlayer();
 		BendingPlayer bPlayer = BendingPlayer.getBendingPlayer(player);
 		if (bPlayer == null) {
@@ -1674,8 +1637,6 @@ public class PKListener implements Listener {
 
 	@EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
 	public void onPlayerToggleFlight(PlayerToggleFlightEvent event) {
-		
-
 		Player player = event.getPlayer();
 		if (CoreAbility.hasAbility(player, Tornado.class) || Bloodbending.isBloodbent(player) || Suffocate.isBreathbent(player) || CoreAbility.hasAbility(player, FireJet.class) || CoreAbility.hasAbility(player, AvatarState.class)) {
 			event.setCancelled(player.getGameMode() != GameMode.CREATIVE);
@@ -1692,6 +1653,7 @@ public class PKListener implements Listener {
 			for (Entity en : GeneralMethods.getEntitiesAroundPoint(loc, smokescreen.getRadius())) {
 				smokescreen.applyBlindness(en);
 			}
+			
 			Smokescreen.getSnowballs().remove(id);
 		}
 	}
