+++ src/com/projectkorra/projectkorra/ability/EarthAbility.java
@@ -25,7 +25,7 @@ import com.projectkorra.projectkorra.ability.util.Collision;
 import com.projectkorra.projectkorra.configuration.ConfigManager;
 import com.projectkorra.projectkorra.earthbending.RaiseEarth;
 import com.projectkorra.projectkorra.earthbending.lava.LavaFlow;
-import com.projectkorra.projectkorra.earthbending.passive.EarthPassive;
+import com.projectkorra.projectkorra.earthbending.passive.DensityShift;
 import com.projectkorra.projectkorra.firebending.Illumination;
 import com.projectkorra.projectkorra.util.BlockSource;
 import com.projectkorra.projectkorra.util.Information;
@@ -152,8 +152,8 @@ public abstract class EarthAbility extends ElementalAbility {
 			}
 
 			Block affectedblock = location.clone().add(norm).getBlock();
-			if (EarthPassive.isPassiveSand(block)) {
-				EarthPassive.revertSand(block);
+			if (DensityShift.isPassiveSand(block)) {
+				DensityShift.revertSand(block);
 			}
 			if (Illumination.isIlluminationTorch(affectedblock) && TempBlock.isTempBlock(affectedblock)) {
 				TempBlock.get(affectedblock).revertBlock();
@@ -204,8 +204,8 @@ public abstract class EarthAbility extends ElementalAbility {
 						}
 						break;
 					}
-					if (EarthPassive.isPassiveSand(affectedblock)) {
-						EarthPassive.revertSand(affectedblock);
+					if (DensityShift.isPassiveSand(affectedblock)) {
+						DensityShift.revertSand(affectedblock);
 					}
 					if (block == null) {
 						for (Block checkblock : blocks) {
@@ -670,7 +670,7 @@ public abstract class EarthAbility extends ElementalAbility {
 	}
 
 	public static void stopBending() {
-		EarthPassive.removeAll();
+		DensityShift.removeAll();
 
 		if (isEarthRevertOn()) {
 			removeAllEarthbendedBlocks();
