+++ src/com/projectkorra/projectkorra/chiblocking/passive/ChiPassive.java
@@ -17,7 +17,6 @@ import com.projectkorra.projectkorra.configuration.ConfigManager;
 import com.projectkorra.projectkorra.util.ActionBar;
 
 public class ChiPassive {
-
 	public static boolean willChiBlock(Player attacker, Player player) {
 		BendingPlayer bPlayer = BendingPlayer.getBendingPlayer(player);
 		if (bPlayer == null) {
@@ -44,6 +43,7 @@ public class ChiPassive {
 		} else if (bPlayer.isChiBlocked()) {
 			return false;
 		}
+		
 		return true;
 	}
 
@@ -56,6 +56,7 @@ public class ChiPassive {
 		if (bPlayer == null) {
 			return;
 		}
+		
 		bPlayer.blockChi();
 		player.getWorld().playSound(player.getLocation(), Sound.ENTITY_ENDERDRAGON_HURT, 2, 0);
 		
@@ -71,23 +72,7 @@ public class ChiPassive {
 			}
 		}.runTaskTimer(ProjectKorra.plugin, 0, 1);
 	}
-
-	public static double getExhaustionFactor() {
-		return ConfigManager.getConfig().getDouble("Abilities.Chi.Passive.ChiSaturation.ExhaustionFactor");
-	}
-
-	public static double getFallReductionFactor() {
-		return ConfigManager.getConfig().getDouble("Abilities.Chi.Passive.Acrobatics.FallReductionFactor");
-	}
-
-	public static int getJumpPower() {
-		return ConfigManager.getConfig().getInt("Abilities.Chi.Passive.ChiAgility.JumpPower");
-	}
-
-	public static int getSpeedPower() {
-		return ConfigManager.getConfig().getInt("Abilities.Chi.Passive.ChiAgility.SpeedPower");
-	}
-
+	
 	public static double getChance() {
 		return ConfigManager.getConfig().getDouble("Abilities.Chi.Passive.BlockChi.Chance");
 	}
@@ -99,5 +84,4 @@ public class ChiPassive {
 	public static long getTicks() {
 		return (getDuration() / 1000) * 20;
 	}
-
 }
