+++ src/com/projectkorra/projectkorra/chiblocking/passive/ChiSaturation.java
@@ -5,17 +5,19 @@ import org.bukkit.entity.Player;
 
 import com.projectkorra.projectkorra.ability.ChiAbility;
 import com.projectkorra.projectkorra.ability.PassiveAbility;
+import com.projectkorra.projectkorra.configuration.ConfigManager;
 
 public class ChiSaturation extends ChiAbility implements PassiveAbility {
-
 	public ChiSaturation(Player player) {
 		super(player);
 	}
 
-	@Override
-	public void progress() {
-
+	public static double getExhaustionFactor() {
+		return ConfigManager.getConfig().getDouble("Abilities.Chi.Passive.ChiSaturation.ExhaustionFactor");
 	}
+
+	@Override
+	public void progress() {}
 
 	@Override
 	public boolean isSneakAbility() {
@@ -39,12 +41,11 @@ public class ChiSaturation extends ChiAbility implements PassiveAbility {
 
 	@Override
 	public Location getLocation() {
-		return null;
+		return player.getLocation();
 	}
 
 	@Override
 	public boolean isInstantiable() {
 		return false;
 	}
-
 }
