+++ src/com/projectkorra/projectkorra/earthbending/util/EarthbendingManager.java
@@ -4,13 +4,12 @@ import com.projectkorra.projectkorra.ProjectKorra;
 import com.projectkorra.projectkorra.earthbending.EarthTunnel;
 import com.projectkorra.projectkorra.earthbending.Shockwave;
 import com.projectkorra.projectkorra.earthbending.Tremorsense;
-import com.projectkorra.projectkorra.earthbending.passive.EarthPassive;
+import com.projectkorra.projectkorra.earthbending.passive.DensityShift;
 import com.projectkorra.projectkorra.util.RevertChecker;
 
 import org.bukkit.Bukkit;
 
 public class EarthbendingManager implements Runnable {
-
 	public ProjectKorra plugin;
 
 	public EarthbendingManager(ProjectKorra plugin) {
@@ -18,8 +17,7 @@ public class EarthbendingManager implements Runnable {
 	}
 
 	public void run() {
-		EarthPassive.revertSands();
-		//EarthPassive.handleMetalPassives();
+		DensityShift.revertSands();
 		RevertChecker.revertEarthBlocks();
 		Shockwave.progressAll();
 		Tremorsense.manage(Bukkit.getServer());
