--- src/com/projectkorra/projectkorra/waterbending/passive/Hydrosink.java
@@ -1,50 +1,0 @@
-package com.projectkorra.projectkorra.waterbending.passive;
-
-import org.bukkit.Location;
-import org.bukkit.entity.Player;
-
-import com.projectkorra.projectkorra.ability.PassiveAbility;
-import com.projectkorra.projectkorra.ability.WaterAbility;
-
-public class Hydrosink extends WaterAbility implements PassiveAbility {
-
-	public Hydrosink(Player player) {
-		super(player);
-	}
-
-	@Override
-	public void progress() {
-
-	}
-
-	@Override
-	public boolean isSneakAbility() {
-		return false;
-	}
-
-	@Override
-	public boolean isHarmlessAbility() {
-		return true;
-	}
-
-	@Override
-	public long getCooldown() {
-		return 0;
-	}
-
-	@Override
-	public String getName() {
-		return "Hydrosink";
-	}
-
-	@Override
-	public Location getLocation() {
-		return null;
-	}
-
-	@Override
-	public boolean isInstantiable() {
-		return false;
-	}
-
-}
